fix: recoil reset 위치 수정
